% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPlibrary3.R
\name{SIPlibrary3}
\alias{SIPlibrary3}
\title{SIPLibrary 3.0 Generator}
\usage{
SIPlibrary3(
  DATAFRAME,
  filename,
  author,
  METADF = NULL,
  seeds = "generate",
  bounds = c(0, 1),
  boundedness = "u",
  term_saved = 5
)
}
\arguments{
\item{DATAFRAME}{Data.frame containing data or SIPs.}

\item{filename}{Character, ending in ".xlsx"}

\item{author}{Character.}

\item{METADF}{Data.frame with named rows of same length as DATAFRAME.}

\item{seeds}{Optional data.frame input for the HDR seeds. Default is "generate",
which calls the HDRSeeds function to generate statistically independent seeds.}

\item{bounds}{Passed to metalog function, numeric vector specifying lower or
upper bounds, none required if the distribution is unbounded.}

\item{boundedness}{Passed to metalog function, character string specifying
unbounded, semi-bounded upper, semi-bounded lower or bounded; accepts values
u, su, sl and b (default: 'u').}

\item{term_saved}{Passed to metalog function, a value between 3 and 16 for the
number of terms desired for all SIPs. Default is 5.}
}
\value{
Completion notification, file output to WD
}
\description{
Takes a data.frame and an optional data.frame containing metadata and coverts
it to a .xlsx format PM.org 3.0 Standard Library readable by ChanceCalc.
Allows a user to easily output R SIPs to Excel users.
}
\examples{
testdf <- data.frame(SIP_1 = rnorm(5000),
SIP_2 = rnorm(5000),
SIP_3 = rnorm(5000),
SIP_4 = rnorm(5000),
SIP_5 = rnorm(5000))

testmeta <- sapply(testdf, FUN = summary)

SIPlibrary3(testdf, "finaltesting.xlsx", "Aaron Brown", testmeta)

}
